requirements.txt
# ============================================================
# SYUTAINβ V25 Python依存関係
# pip install -r requirements.txt
# ============================================================

# ===== Web フレームワーク =====
fastapi>=0.115
uvicorn[standard]>=0.34
sse-starlette>=2.3
websockets>=15.0
python-multipart>=0.0.20
pydantic>=2.11
pydantic-settings>=2.9

# ===== データベース =====
asyncpg>=0.30
psycopg2-binary>=2.9
sqlalchemy[asyncio]>=2.0
alembic>=1.15
aiosqlite>=0.21

# ===== NATS =====
nats-py>=2.10

# ===== LLM クライアント =====
openai>=1.82
anthropic>=0.52
google-genai>=1.16
httpx>=0.28

# ===== 情報収集 =====
tavily-python>=0.7
feedparser>=6.0
google-api-python-client>=2.170
google-auth-oauthlib>=1.2

# ===== MCP =====
mcp>=1.9

# ===== SNS =====
tweepy>=4.15

# ===== 暗号通貨 =====
ccxt>=4.4

# ===== ユーティリティ =====
python-dotenv>=1.1
pyyaml>=6.0
jinja2>=3.1
python-jose[cryptography]>=3.4
passlib[bcrypt]>=1.7
aiofiles>=24.1
tenacity>=9.1
structlog>=25.4

# ===== 監視・メトリクス =====
psutil>=7.0

# ===== ベクトル =====
numpy>=2.2

# ===== 通知 =====
aiohttp>=3.12

# ===== スケジューラ =====
apscheduler>=3.11
